Create modules in the data store if not exist
