+++ src/CompositionRoot.ts
@@ -6,6 +6,7 @@ import { GetCurrentUserUseCase } from "./domain/usecases/GetCurrentUserUseCase";
 import { GetGlassModuleByNameUseCase } from "./domain/usecases/GetGlassModuleByNameUseCase";
 import { GetGlassModulesUseCase } from "./domain/usecases/GetGlassModulesUseCase";
 import { GetInstanceVersionUseCase } from "./domain/usecases/GetInstanceVersionUseCase";
+import { ValidateGlassModulesUseCase } from "./domain/usecases/ValidateGlassModulesUseCase";
 
 export function getCompositionRoot(instance: Instance) {
     const dataStoreClient = new DataStoreClient(instance);
@@ -20,6 +21,7 @@ export function getCompositionRoot(instance: Instance) {
         glassModules: getExecute({
             getAll: new GetGlassModulesUseCase(glassModuleRepository),
             getByName: new GetGlassModuleByNameUseCase(glassModuleRepository),
+            validate: new ValidateGlassModulesUseCase(glassModuleRepository),
         }),
     };
 }
