+++ src/domain/repositories/GlassModuleRepository.ts
@@ -4,4 +4,5 @@ import { GlassModule } from "../entities/GlassModule";
 export interface GlassModuleRepository {
     getAll(): FutureData<GlassModule[]>;
     getByName(name: string): FutureData<GlassModule>;
+    save(modules: GlassModule[]): FutureData<void>;
 }
